feat(hook-reverse-proxy): add hot relead config
